requirements.txt
Django
pymongo[snappy,gssapi,srv,tls]
dnspython
djangorestframework
django-cors-headers